Move allegro5.rb to allegro.rb and make some improvements.
